Flex.Worker module
