+++ lib/flex.ex
@@ -6,41 +6,6 @@ defmodule Flex do
   @sec 1_000
 
   @doc """
-  Given a string representing a .flac file, create a corresponding .mp3 file for the .flac file.
-  """
-  def convert_flac(flacfile) do
-    dirname = Path.dirname(flacfile)
-    basename = Path.basename(flacfile, ".flac")
-    wavfile = Path.join(dirname, "#{basename}.wav")
-    mp3file = Path.join(dirname, "#{basename}.mp3")
-
-    flac_to_wav = fn ->
-      System.cmd("flac", ["--silent", "--force", "--decode", "--output-name", wavfile, flacfile], stderr_to_stdout: false)
-    end
-
-    wav_to_mp3 = fn ->
-      System.cmd("lame", ["--silent", "--abr", "320", wavfile, mp3file], stderr_to_stdout: false)
-    end
-
-    remove_wav = fn ->
-      System.cmd("rm", [wavfile])
-    end
-
-    IO.write "starting on #{basename}..."
-
-    Task.async(flac_to_wav)
-    |> Task.await 10*@sec
-
-    Task.async(wav_to_mp3)
-    |> Task.await 30*@sec
-
-    Task.async(remove_wav)
-    |> Task.await 1*@sec
-
-    IO.puts " done"
-  end
-
-  @doc """
   Given a string representing a directory, kick off conversions for each .flac file in the directory.
   """
   def convert_dir(dir\\".") do
@@ -49,7 +14,7 @@ defmodule Flex do
     Path.expand(dir)
     |> Path.join("**/*.flac")
     |> Path.wildcard
-    |> Stream.map(&(Task.async(fn -> convert_flac(&1) end)))
+    |> Stream.map(&(Task.async(fn -> Flex.Worker.convert_flac(&1) end)))
     |> Enum.map(&(Task.await &1, 50 * @sec))
   end
 
@@ -72,4 +37,3 @@ defmodule Flex do
   end
 
 end
-
